Refactor message component Standard naming between list and full modes
