+++ src/components/message/Full.vue
